cleaned up version of up/downhill parametrization with new formulas
