+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/CostCalculatorMTB.java
@@ -9,20 +9,20 @@ import mg.mgmap.generic.graph.WayAttributs;
 
 public class CostCalculatorMTB implements CostCalculator {
 
-    private static final double ful = 1 - 0.833;
-    private static final double fdl = 1 - 0.833;
-    protected double mUpCosts;
+    private static final double ful = 1 - 0.833; // reduction of uphill slope limit with delta mtbscaleUp
+    private static final double fdl = 1 - 0.833; // reduction of downhill slope limit with delta mtbscale
+/*    protected double mUpCosts;
     protected double mDnCosts;// base costs in m per hm uphill;
-    protected double mUpSlopeLimit; //  up to this slope base Costs
-    protected double mDnSlopeLimit;
     protected double mUpAddCosts; // relative additional costs per slope increase ( 10 means, that costs double with 10% slope increase )
-    protected double mDnAddCosts;
+    protected double mDnAddCosts; */
+    private final double mUpSlopeLimit; //  up to this slope base Costs
+    private final double mDnSlopeLimit;
+    private final double mDelta;
+    private final double mfud;
+    private final double mfdd;
 
-    double mfud;
-    double mfdd;
-
-    private CostCalculatorHeuristicTwoPieceFunc mProfileCalculator;
-    public CostCalculatorMTB(WayAttributs wayTagEval, CostCalculatorHeuristicTwoPieceFunc profile,  double pful, double base_ul ,double base_us,double tful, double base_dl ,double base_ds) {
+    private CostCalculatorTwoPieceFunc mProfileCalculator;
+    public CostCalculatorMTB(WayAttributs wayTagEval, CostCalculatorTwoPieceFunc profile) {
         mProfileCalculator = profile;
         double mtbUp = -2;
         double mtbDn = -2;
@@ -59,16 +59,41 @@ public class CostCalculatorMTB implements CostCalculator {
         }
 
         mUpSlopeLimit = profile.mUpSlopeLimit * (1-ful*deltaUp);
-        mUpAddCosts = mProfileCalculator.fus/( mUpSlopeLimit * mUpSlopeLimit);
-        mUpCosts =  mProfileCalculator.fu /mUpSlopeLimit;
-
-        mDnSlopeLimit = profile.mDnSlopeLimit * (1-fdl*deltaDn);
-        mDnAddCosts = mProfileCalculator.fds/( mDnSlopeLimit * mDnSlopeLimit) ;
-        mDnCosts = mProfileCalculator.mDnCosts ;// base costs in m per hm uphill;
-
+/*        mUpAddCosts = mProfileCalculator.fus/( mUpSlopeLimit * mUpSlopeLimit);
+        mUpCosts =  mProfileCalculator.fu /mUpSlopeLimit; //
+*/
+        mDelta =  (1-fdl*deltaDn);
+        mDnSlopeLimit = profile.mDnSlopeLimit * mDelta;
+/*        mDnAddCosts = mProfileCalculator.fds/( mDnSlopeLimit * mDnSlopeLimit) ;
+        mDnCosts = mProfileCalculator.mDnCosts ;// attention! Need to be different approach than upCosts to avoid conflict with heuristic
+*/
     }
 
     public double calcCosts(double dist, float vertDist){
+        if (dist <= 0.0000001 ) {
+            return 0.0001;
+        }
+        double slope = vertDist / dist;
+        if ( abs(slope) >=  10.0 ) Log.e("CostCalcMTB","Suspicious Slope in calcCosts. Dist:" + dist + " VertDist:" + vertDist + " Slope:" + slope);
+        double cost;
+        double relslope;
+        if (slope >= 0) {
+            relslope = slope / mUpSlopeLimit;
+            if (relslope <= 1)
+                cost = dist* ( mfud + relslope * mProfileCalculator.fu);
+            else
+                cost = dist* ( mfud + relslope * ( mProfileCalculator.fu + (relslope - 1) * mProfileCalculator.fus));
+        } else {
+            relslope = slope / mDnSlopeLimit;
+            if (relslope <= 1)
+                cost = dist* ( mfdd + relslope * mDelta * mProfileCalculator.fd);
+            else
+                cost = dist* ( mfdd + relslope * mDelta *( mProfileCalculator.fd + (relslope - 1) * mProfileCalculator.fds));
+        }
+        return cost + 0.0001;
+    }
+
+/*    public double calcCostsOld(double dist, float vertDist){
         if (dist <= 0.0000001 ) {
             return 0.0001;
         }
@@ -84,10 +109,11 @@ public class CostCalculatorMTB implements CostCalculator {
             if (slope >= mDnSlopeLimit)
                 cost = dist* mfdd + vertDist * mDnCosts;
             else
-                return dist* mfdd + vertDist * ( mDnCosts + (slope - mDnSlopeLimit) * mDnAddCosts);
+                cost = dist* mfdd + vertDist * ( mDnCosts + (slope - mDnSlopeLimit) * mDnAddCosts);
         }
         return cost + 0.0001;
-    }
+    } */
+
 
     @Override
     public double heuristic(double dist, float vertDist) {
