+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/MTB_K1S1.java
@@ -7,11 +7,11 @@ import mg.mgmap.generic.graph.WayAttributs;
 
 public class MTB_K1S1 extends RoutingProfile {
     public MTB_K1S1( ) {
-        super(new CostCalculatorHeuristicTwoPieceFunc( 1,  1, -0.27, 2));
+        super(new CostCalculatorTwoPieceFunc(1,  1));
     }
 
     protected CostCalculator getCostCalculator(CostCalculator profileCalculator, WayAttributs wayAttributs) {
-        return new CostCalculatorMTB(wayAttributs, (CostCalculatorHeuristicTwoPieceFunc) profileCalculator, 1.1,0.0,1.4,1.0,0.8,0.5);
+        return new CostCalculatorMTB(wayAttributs, (CostCalculatorTwoPieceFunc) profileCalculator);
     }
