+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/MTB_K1S3.java
@@ -7,11 +7,11 @@ import mg.mgmap.generic.graph.WayAttributs;
 
 public class MTB_K1S3 extends RoutingProfile {
     public MTB_K1S3( ) {
-        super(new CostCalculatorHeuristicTwoPieceFunc( 1,  3, -0.27, 2));
+        super(new CostCalculatorTwoPieceFunc( 1,  3));
     }
 
     protected CostCalculator getCostCalculator(CostCalculator profileCalculator, WayAttributs wayAttributs) {
-        return new CostCalculatorMTB(wayAttributs, (CostCalculatorHeuristicTwoPieceFunc) profileCalculator, 1.1,0.0,2.2,1.0,0.8,1.5); //0.0,0.0,0.7,0.0,0.7,0.0);
+        return new CostCalculatorMTB(wayAttributs, (CostCalculatorTwoPieceFunc) profileCalculator); //0.0,0.0,0.7,0.0,0.7,0.0);
     }
 
     @Override
