+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/MTB_K2S2.java
@@ -7,12 +7,12 @@ import mg.mgmap.generic.graph.WayAttributs;
 
 public class MTB_K2S2 extends RoutingProfile {
     public MTB_K2S2( ) {
-        super(new CostCalculatorHeuristicTwoPieceFunc(2.0,  2, -0.27, 2));
+        super(new CostCalculatorTwoPieceFunc(2.0,  2));
     }
 
 
     protected CostCalculator getCostCalculator(CostCalculator profileCalculator, WayAttributs wayAttributs) {
-        return new CostCalculatorMTB(wayAttributs, (CostCalculatorHeuristicTwoPieceFunc) profileCalculator, 1.2,0.0,1.8,1.0,0.8,1.0); //0.0,0.0,0.7,0.0,0.7,0.0);
+        return new CostCalculatorMTB(wayAttributs, (CostCalculatorTwoPieceFunc) profileCalculator); //0.0,0.0,0.7,0.0,0.7,0.0);
     }
